+++ components/form/builder/src/Standard/Localization/EsEsLocalization.js
@@ -1,0 +1,11 @@
+import {Localization} from './Localization'
+
+export class EsEsLocalization extends Localization {
+  fromStringToLocaleFloat() {
+    return parseFloat(
+      this.removeAllCommas(
+        this.replaceFirstCommaWithDot(this.removeAllDots(this._value))
+      )
+    )
+  }
+}